refacor: 레이싱 진행 부분, 레이싱 설정 부분 분리
